[fix] fixed all warning (mostly float to int conversion)

+++ EngineArchitecture/Source/Engine/GameTool/Component.cpp
@@ -1,7 +1,7 @@
 #include "Component.h"
 #include "Actor.h"
 
-Component::Component(float pUpdateOrder) : mUpdateOrder(pUpdateOrder)
+Component::Component(int pUpdateOrder) : mUpdateOrder(pUpdateOrder)
 {
 }
